usb: Fix possible errors in msm_hsusb_host
PVS-Studio:
V547 Expression 'hcd->irq < 0' is always false. Unsigned type value is never < 0

+++ drivers/usb/host/ehci-msm72k.c
@@ -682,7 +682,7 @@ static int __devinit ehci_msm_probe(struct platform_device *pdev)
 		return  -ENOMEM;
 
 	hcd->irq = platform_get_irq(pdev, 0);
-	if (hcd->irq < 0) {
+	if ((int)hcd->irq < 0) {
 		usb_put_hcd(hcd);
 		return hcd->irq;
 	}
